+++ commitlint.config.js
@@ -16,4 +16,13 @@ limitations under the License.
 
 */
 
-module.exports = { extends: ['@commitlint/config-conventional'] }
+module.exports = {
+  extends: ['@commitlint/config-conventional'],
+  rules: {
+    // Casing: Use whatever casing you'd like!
+    'header-case': [0],
+    'scope-case': [0],
+    'subject-case': [0],
+    'header-max-length': [0],
+  },
+}