+++ packages/@tinacms/icons/src/ChevronRight.tsx
@@ -16,4 +16,16 @@ limitations under the License.
 
 */
 
-module.exports = require('../../typedoc.js')(require('./package.json'))
+import * as React from 'react'
+
+export const ChevronRightIcon = () => (
+  <svg
+    width="32"
+    height="32"
+    viewBox="0 0 32 32"
+    fill="inherit"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path d="M11 24.792L12.2654 26L21.4773 17.2061C22.1747 16.5403 22.1737 15.4588 21.4773 14.7939L12.2654 6L11 7.208L20.2099 16L11 24.792Z" />
+  </svg>
+)