+++ packages/react-tinacms-inline/src/blocks/add-block-menu.tsx
@@ -19,18 +19,46 @@ limitations under the License.
 import * as React from 'react'
 import styled, { css } from 'styled-components'
 import { BlockTemplate } from 'tinacms'
-import { Button } from '@tinacms/styles'
+import { IconButton } from '@tinacms/styles'
 import { AddIcon } from '@tinacms/icons'
 
 interface AddBlockMenuProps {
   addBlock(data: any): void
   templates: BlockTemplate[]
+  position?: 'top' | 'bottom' | 'left' | 'right'
+  index?: number
+  offset?: number
 }
 
-export function AddBlockMenu({ templates, addBlock }: AddBlockMenuProps) {
+export function AddBlockMenu({
+  templates,
+  addBlock,
+  position,
+  index,
+  offset,
+}: AddBlockMenuProps) {
   const [isOpen, setIsOpen] = React.useState(false)
+  const addBlockButtonRef = React.useRef<HTMLButtonElement>(null)
+  const [openTop, setOpenTop] = React.useState(false)
 
-  const handleOpenBlockMenu = () => {
+  const handleOpenBlockMenu = (event: any) => {
+    event.stopPropagation()
+    event.preventDefault()
+    const addBlockButtonElem = addBlockButtonRef.current
+
+    if (addBlockButtonElem !== null) {
+      const menuBounding = addBlockButtonElem.getBoundingClientRect()
+      const halfWindowHeight =
+        (window.innerHeight || document.documentElement.clientHeight) / 2
+      const offsetTop = menuBounding.top - window.scrollY
+
+      if (offsetTop < halfWindowHeight) {
+        setOpenTop(false)
+      } else {
+        setOpenTop(true)
+      }
+    }
+
     templates.length == 1
       ? addBlock({
           _template: templates[0].type,
@@ -48,15 +76,28 @@ export function AddBlockMenu({ templates, addBlock }: AddBlockMenuProps) {
   templates = templates || []
 
   return (
-    <AddBlockWrapper>
-      <AddBlockButton onClick={handleOpenBlockMenu} isOpen={isOpen} primary>
-        <AddIcon /> Add Block
+    <AddBlockWrapper
+      index={index}
+      offset={offset}
+      position={position}
+      isOpen={isOpen}
+    >
+      <AddBlockButton
+        ref={addBlockButtonRef}
+        onClick={handleOpenBlockMenu}
+        isOpen={isOpen}
+        primary
+        small
+      >
+        <AddIcon />
       </AddBlockButton>
-      <BlocksMenu isOpen={isOpen}>
+      <BlocksMenu openTop={openTop} isOpen={isOpen}>
         {templates.map((template: BlockTemplate) => (
           <BlockOption
             key={template.label}
-            onClick={() => {
+            onClick={event => {
+              event.stopPropagation()
+              event.preventDefault()
               addBlock({
                 _template: template.type,
                 ...template.defaultItem,
@@ -74,9 +115,10 @@ export function AddBlockMenu({ templates, addBlock }: AddBlockMenuProps) {
 interface AddMenuProps {
   isOpen?: boolean
   active?: boolean
+  openTop?: boolean
 }
 
-const AddBlockButton = styled(Button)<AddMenuProps>`
+const AddBlockButton = styled(IconButton)<AddMenuProps>`
   font-family: 'Inter', sans-serif;
   display: flex;
   align-items: center;
@@ -86,14 +128,11 @@ const AddBlockButton = styled(Button)<AddMenuProps>`
   }
 
   svg {
-    height: 70%;
-    width: auto;
-    margin-right: 0.5em;
     transition: all 150ms ease-out;
   }
 
   ${props =>
-    props.open &&
+    props.isOpen &&
     css`
       svg {
         transform: rotate(45deg);
@@ -101,20 +140,76 @@ const AddBlockButton = styled(Button)<AddMenuProps>`
     `};
 `
 
-const AddBlockWrapper = styled.div`
-  position: relative;
+interface AddBlockWrapperProps {
+  index?: number
+  offset?: number
+  position?: 'top' | 'bottom' | 'left' | 'right'
+  isOpen: boolean
+}
+
+const AddBlockWrapper = styled.div<AddBlockWrapperProps>(
+  p => css`
+  position: absolute;
+  z-index: calc(var(--tina-z-index-1) - ${p.index !== undefined ? p.index : 0});
+
+  ${p.position == 'top' &&
+    css`
+      top: -${p.offset !== undefined ? p.offset : `16`}px;
+      left: 50%;
+      transform: translate3d(-50%, -50%, 0);
+    `}
+
+  ${p.position == 'left' &&
+    css`
+      top: 50%;
+      left: -${p.offset !== undefined ? p.offset : `16`}px;
+      transform: translate3d(-50%, -50%, 0);
+    `}
+
+  ${p.position == 'bottom' &&
+    css`
+      bottom: -${p.offset !== undefined ? p.offset : `16`}px;
+      left: 50%;
+      transform: translate3d(-50%, 50%, 0);
+    `}
+
+  ${p.position == 'right' &&
+    css`
+      top: 50%;
+      right: -${p.offset !== undefined ? p.offset : `16`}px;
+      transform: translate3d(50%, -50%, 0);
+    `}
+
+  ${p.position == undefined &&
+    css`
+      position: relative;
+      left: 0;
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      width: 100%;
+    `}
+
+  ${p.isOpen &&
+    css`
+      z-index: calc(1 + var(--tina-z-index-1) - ${p.index ? p.index : 0});
+    `}
 `
+)
 
 const BlocksMenu = styled.div<AddMenuProps>`
   min-width: 192px;
   border-radius: var(--tina-radius-big);
   border: 1px solid var(--tina-color-grey-2);
-  display: block;
+  display: flex;
+  flex-direction: column;
+  align-items: stretch;
+  justify-content: flex-start;
   position: absolute;
   z-index: var(--tina-z-index-2);
-  top: 0;
-  left: 0;
-  transform: translate3d(0, 0, 0) scale3d(0.5, 0.5, 1);
+  top: 4px;
+  left: 50%;
+  transform: translate3d(-50%, 0, 0) scale3d(0.5, 0.5, 1);
   opacity: 0;
   pointer-events: none;
   transition: all 150ms ease-out;
@@ -128,7 +223,20 @@ const BlocksMenu = styled.div<AddMenuProps>`
     css`
       opacity: 1;
       pointer-events: all;
-      transform: translate3d(0, 41px, 0) scale3d(1, 1, 1);
+      transform: translate3d(-50%, 32px, 0) scale3d(1, 1, 1);
+    `};
+
+  ${props =>
+    props.openTop &&
+    css`
+      top: auto;
+      bottom: 4px;
+      transform-origin: 50% 100%;
+
+      ${props.isOpen &&
+        css`
+          transform: translate3d(-50%, -32px, 0) scale3d(1, 1, 1);
+        `};
     `};
 `
 
@@ -139,7 +247,7 @@ const BlockOption = styled.button`
   font-size: var(--tina-font-size-1);
   padding: 0 12px;
   height: 34px;
-  font-weight: 500;
+  font-weight: var(--tina-font-weight-regular);
   width: 100%;
   background: none;
   cursor: pointer;