+++ packages/react-tinacms-inline/src/blocks/inline-block-text.tsx
@@ -17,18 +17,11 @@ limitations under the License.
 */
 
 import * as React from 'react'
-import { BlockField } from './inline-block-field'
-import { InlineTextFieldProps, InlineText } from '../inline-field-text'
+import { InlineTextProps, InlineText } from '../fields/inline-text-field'
 
-export function BlockText({ name, className }: InlineTextFieldProps) {
-  return (
-    <BlockField name={name}>
-      {({ input, status }) => {
-        if (status === 'active') {
-          return <InlineText {...input} className={className} />
-        }
-        return <>{input.value}</>
-      }}
-    </BlockField>
-  )
+/**
+ * @deprecated
+ */
+export function BlockText(props: InlineTextProps) {
+  return <InlineText {...props} focusRing={false} />
 }