+++ packages/@tinacms/react-modals/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.7](https://github.com/tinacms/tinacms/compare/@tinacms/react-modals@0.1.7-alpha.2...@tinacms/react-modals@0.1.7) (2020-06-01)
+
+**Note:** Version bump only for package @tinacms/react-modals
+
+
+
+
+
+## [0.1.7-alpha.2](https://github.com/tinacms/tinacms/compare/@tinacms/react-modals@0.1.7-alpha.1...@tinacms/react-modals@0.1.7-alpha.2) (2020-06-01)
+
+**Note:** Version bump only for package @tinacms/react-modals
+
+
+
+
+
+## [0.1.7-alpha.1](https://github.com/tinacms/tinacms/compare/@tinacms/react-modals@0.1.7-alpha.0...@tinacms/react-modals@0.1.7-alpha.1) (2020-05-29)
+
+**Note:** Version bump only for package @tinacms/react-modals
+
+
+
+
+
+## [0.1.7-alpha.0](https://github.com/tinacms/tinacms/compare/@tinacms/react-modals@0.1.6...@tinacms/react-modals@0.1.7-alpha.0) (2020-05-28)
+
+**Note:** Version bump only for package @tinacms/react-modals
+
+
+
+
+
+## [0.1.6](https://github.com/tinacms/tinacms/compare/@tinacms/react-modals@0.1.5...@tinacms/react-modals@0.1.6) (2020-05-25)
+
+**Note:** Version bump only for package @tinacms/react-modals
+
+
+
+
+
 ## [0.1.5](https://github.com/tinacms/tinacms/compare/@tinacms/react-modals@0.1.4...@tinacms/react-modals@0.1.5) (2020-05-19)
 
 **Note:** Version bump only for package @tinacms/react-modals