+++ packages/@tinacms/styles/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.5.0](https://github.com/tinacms/tinacms/compare/@tinacms/styles@0.5.0-alpha.0...@tinacms/styles@0.5.0) (2020-06-01)
+
+**Note:** Version bump only for package @tinacms/styles
+
+
+
+
+
+# [0.5.0-alpha.0](https://github.com/tinacms/tinacms/compare/@tinacms/styles@0.4.2...@tinacms/styles@0.5.0-alpha.0) (2020-05-28)
+
+
+### Features
+
+* add font loader ([4f37605](https://github.com/tinacms/tinacms/commit/4f37605))
+* tina provider accepts 'styled' prop ([c581595](https://github.com/tinacms/tinacms/commit/c581595))
+
+
+
+
+
 ## [0.4.2](https://github.com/tinacms/tinacms/compare/@tinacms/styles@0.4.2-alpha.2...@tinacms/styles@0.4.2) (2020-04-27)
 
 **Note:** Version bump only for package @tinacms/styles