+++ packages/demo-cra/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.33](https://github.com/tinacms/tinacms/compare/demo-cra@1.0.33-alpha.2...demo-cra@1.0.33) (2020-06-01)
+
+**Note:** Version bump only for package demo-cra
+
+
+
+
+
+## [1.0.33-alpha.2](https://github.com/tinacms/tinacms/compare/demo-cra@1.0.33-alpha.1...demo-cra@1.0.33-alpha.2) (2020-06-01)
+
+**Note:** Version bump only for package demo-cra
+
+
+
+
+
+## [1.0.33-alpha.1](https://github.com/tinacms/tinacms/compare/demo-cra@1.0.33-alpha.0...demo-cra@1.0.33-alpha.1) (2020-05-29)
+
+**Note:** Version bump only for package demo-cra
+
+
+
+
+
+## [1.0.33-alpha.0](https://github.com/tinacms/tinacms/compare/demo-cra@1.0.32...demo-cra@1.0.33-alpha.0) (2020-05-28)
+
+**Note:** Version bump only for package demo-cra
+
+
+
+
+
+## [1.0.32](https://github.com/tinacms/tinacms/compare/demo-cra@1.0.31...demo-cra@1.0.32) (2020-05-25)
+
+**Note:** Version bump only for package demo-cra
+
+
+
+
+
 ## [1.0.31](https://github.com/tinacms/tinacms/compare/demo-cra@1.0.30...demo-cra@1.0.31) (2020-05-19)
 
 **Note:** Version bump only for package demo-cra