+++ packages/gatsby-tinacms-json/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.25](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-json@0.8.25-alpha.2...gatsby-tinacms-json@0.8.25) (2020-06-01)
+
+**Note:** Version bump only for package gatsby-tinacms-json
+
+
+
+
+
+## [0.8.25-alpha.2](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-json@0.8.25-alpha.1...gatsby-tinacms-json@0.8.25-alpha.2) (2020-06-01)
+
+**Note:** Version bump only for package gatsby-tinacms-json
+
+
+
+
+
+## [0.8.25-alpha.1](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-json@0.8.25-alpha.0...gatsby-tinacms-json@0.8.25-alpha.1) (2020-05-29)
+
+**Note:** Version bump only for package gatsby-tinacms-json
+
+
+
+
+
+## [0.8.25-alpha.0](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-json@0.8.24...gatsby-tinacms-json@0.8.25-alpha.0) (2020-05-28)
+
+**Note:** Version bump only for package gatsby-tinacms-json
+
+
+
+
+
+## [0.8.24](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-json@0.8.23...gatsby-tinacms-json@0.8.24) (2020-05-25)
+
+**Note:** Version bump only for package gatsby-tinacms-json
+
+
+
+
+
 ## [0.8.23](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-json@0.8.22...gatsby-tinacms-json@0.8.23) (2020-05-19)
 
 **Note:** Version bump only for package gatsby-tinacms-json